requirements.txt
flask
gunicorn

pandas
xlsxwriter
numpy

python-doctr
pytesseract
opencv-python
ocrmypdf
pillow
pypdf2
torch
